+++ browser-extension/src/features/remove-mobile-css-rules.tsx
@@ -1,0 +1,23 @@
+import features from '@src/feature-registry';
+
+export function init() {
+  for (let i = 0; i < document.styleSheets.length; i++) {
+    const styleSheet = document.styleSheets[i];
+    const rules = styleSheet.cssRules;
+    try {
+      for (let j = rules.length - 1; j >= 0; j--) {
+        const rule = rules[j];
+        if (rule instanceof CSSMediaRule && rule.media.mediaText.includes('screen')) {
+          styleSheet.deleteRule(j);
+        }
+      }
+    } catch (e) {
+      console.log(`Could not modify stylesheet: ${styleSheet.href}, Error: ${e}`);
+    }
+  }
+}
+
+void features.add({
+  id: 'remove-mobile-css-rules',
+  init,
+});